% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_gstat.R
\name{cb_gstat_read}
\alias{cb_gstat_read}
\title{Read GRADISTAT data and save to Excel}
\usage{
cb_gstat_read(path)
}
\arguments{
\item{path}{path to single file or directory}
}
\value{
tibble with cleaned data
}
\description{
Read GRADISTAT data and save to Excel
}
