% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_read.R
\name{cb_read}
\alias{cb_read}
\title{Read and write Beckman Coulter Laser diffraction output spreadsheets}
\usage{
cb_read(path, mode)
}
\arguments{
\item{path}{path to single file or directory}

\item{mode}{processing one file or directory}
}
\value{
a tibble with cleaned data
}
\description{
Read and write Beckman Coulter Laser diffraction output spreadsheets
}
