% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpsa_gstat.R
\name{tpsa_gstat_populate}
\alias{tpsa_gstat_populate}
\title{Populate GRADISTAT file with PSA data}
\usage{
tpsa_gstat_populate(path, gradistat)
}
\arguments{
\item{path}{path to clean PSA data.}

\item{gradistat}{path to empty GRADISTAT file.}
}
\value{
invisible tibble with fitting GRADISTAT data structure.
}
\description{
Populate GRADISTAT file with PSA data
}
