% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpsa_gstat.R
\name{tpsa_gstat_read}
\alias{tpsa_gstat_read}
\title{Read GRADISTAT data and save to Excel file}
\usage{
tpsa_gstat_read(path)
}
\arguments{
\item{path}{path to single file or directory}
}
\value{
tibble with cleaned data
}
\description{
Read GRADISTAT data and save to Excel file
}
