% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpsa_sus.R
\name{tpsa_read_sus}
\alias{tpsa_read_sus}
\title{Read and write Bartington magnetic susceptibility sus files.}
\usage{
tpsa_read_sus(path)
}
\arguments{
\item{path}{path to single file.}
}
\value{
a tibble with cleaned data.
}
\description{
Read and write Bartington magnetic susceptibility sus files.
}
